+++ crates/brontes-types/src/structured_trace.rs
@@ -83,8 +83,8 @@ pub struct TxTrace {
     pub trace:           Vec<TransactionTraceWithLogs>,
     pub state_diff:      StateDiff,
     pub tx_hash:         B256,
-    pub gas_used:        u64,
-    pub effective_price: u64,
+    pub gas_used:        u128,
+    pub effective_price: u128,
     pub tx_index:        u64,
     pub is_success:      bool,
 }
@@ -94,8 +94,8 @@ impl TxTrace {
         trace: Vec<TransactionTraceWithLogs>,
         tx_hash: B256,
         tx_index: u64,
-        gas_used: u64,
-        effective_price: u64,
+        gas_used: u128,
+        effective_price: u128,
         is_success: bool,
         state_diff: StateDiff,
     ) -> Self {
